move api code out
