+++ crates/api/api/Cargo.toml
@@ -26,10 +26,13 @@ lemmy_db_views_comment = { workspace = true, features = ["full"] }
 lemmy_db_views_community = { workspace = true, features = ["full"] }
 lemmy_db_views_community_moderator = { workspace = true, features = ["full"] }
 lemmy_db_views_community_follower = { workspace = true, features = ["full"] }
+lemmy_apub = { workspace = true, features = ["full"] }
+lemmy_apub_objects = { workspace = true, features = ["full"] }
 lemmy_db_views_community_person_ban = { workspace = true, features = ["full"] }
 lemmy_db_views_post = { workspace = true, features = ["full"] }
 lemmy_db_views_vote = { workspace = true, features = ["full"] }
 lemmy_db_views_local_user = { workspace = true, features = ["full"] }
+lemmy_db_views_search_combined = { workspace = true, features = ["full"] }
 lemmy_db_views_person = { workspace = true, features = ["full"] }
 lemmy_db_views_local_image = { workspace = true, features = ["full"] }
 lemmy_db_views_notification = { workspace = true, features = ["full"] }
@@ -64,6 +67,10 @@ sitemap-rs = "0.2.2"
 totp-rs = { version = "5.7.0", features = ["gen_secret", "otpauth"] }
 diesel-async = { workspace = true, features = ["deadpool", "postgres"] }
 either = { workspace = true }
+futures = { workspace = true }
+serde = { workspace = true }
+itertools = { workspace = true }
+serde_json = { workspace = true }
 
 [dev-dependencies]
 serial_test = { workspace = true }
