+++ crates/api/api/src/stuff/read_community.rs
@@ -1,4 +1,4 @@
-use crate::fetcher::resolve_ap_identifier;
+use lemmy_apub::fetcher::resolve_ap_identifier;
 use activitypub_federation::config::Data;
 use actix_web::web::{Json, Query};
 use lemmy_api_utils::{
