+++ crates/apub/src/lib.rs
@@ -16,7 +16,6 @@ use url::Url;
 
 pub mod activities;
 pub mod activity_lists;
-pub mod api;
 pub mod collections;
 pub mod fetcher;
 pub mod http;
